Put all onomastic methods in a single concern
